+++ src/init.c
@@ -53,6 +53,7 @@ static const R_CallMethodDef callMethods[]  = {
 
   { "processx_connection_create",     (DL_FUNC) &processx_connection_create,     2 },
   { "processx_connection_read_chars", (DL_FUNC) &processx_connection_read_chars, 2 },
+  { "processx_connection_read_bytes", (DL_FUNC) &processx_connection_read_bytes, 2 },
   { "processx_connection_read_lines", (DL_FUNC) &processx_connection_read_lines, 2 },
   { "processx_connection_write_bytes",(DL_FUNC) &processx_connection_write_bytes,2 },
   { "processx_connection_file_name",  (DL_FUNC) &processx_connection_file_name,  1 },